Updated to run from anywhere (Linux tested only)
Program modified so when running from terminal in Linux, it will take on the terminal environment to find the source and place the output in the same directory. Also added the option to run silent (-s) or to display progress to the Terminal. Updated to allow search for include and INCLUDE in source code. To run silent, usage: incmerge -s program1.bas program2.bas. To run with progress messages, usage: incmerge program1.bas program2.bas.